// src/types.ts
export enum Role {
  SUPER_ADMIN = 'SUPER_ADMIN',
  PRINCIPAL = 'PRINCIPAL',
  HOD = 'HOD',
  FACULTY = 'FACULTY',
  STAFF = 'STAFF',
  STUDENT = 'STUDENT',
}

export enum Branch {
  CS = 'CS',
  EC = 'EC',
  EEE = 'EEE',
}

export interface College {
  id: string;
  name: string;
  code: string;
  location?: string;
  latitude?: number;
  longitude?: number;
  radius?: number; // in meters
}

export interface User {
  id: string;
  pin: string;
  name: string;
  role: Role;
  branch: string;
  year?: number;
  college_code?: string;
  collegeId?: string;
  email?: string;
  email_verified: boolean;
  parent_email?: string;
  parent_email_verified: boolean;
  phoneNumber?: string;
  imageUrl?: string;
  referenceImageUrl?: string;
  password?: string;
  // New fields for student registration
  fatherName?: string;
  aadharNumber?: string;
  parentPhoneNumber?: string;
  tenthMarks?: string;
  documents?: { name: string; url: string; type: string }[];
}